refactor: 과도한 axios 호출 개선
